refactor: migrate currency from USD to VND across the app

+++ app/src/main/java/com/example/easymart/presentation/ui/admin/products/detail/components/ProductDetailSmallComponents.kt
@@ -162,7 +162,7 @@ fun ProductInfoTile(
     }
 }
 
-fun formatProductPrice(price: Double): String {
+fun formatProductPrice(price: Long): String {
     return "%,.0f đ".format(price)
 }
